# docs/swagger/loyaltysystem-api-docs.yaml
openapi: 3.0.3
info:
  title: Loyalty System API
  description: |
    This is the API documentation for the Loyalty System platform, covering all the endpoints required for:
    * Customer App - Mobile application for end customers
    * Staff App - Application for store staff to handle in-store operations
    * Admin App - Administrative portal for business owners and managers
  version: 1.0.0
servers:
  - url: https://api.loyaltysystem.com/v1
    description: Production server
  - url: https://staging-api.loyaltysystem.com/v1
    description: Staging server
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoyaltyCard:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        loyaltyProgramId:
          type: string
          format: uuid
        cardNumber:
          type: string
        currentPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
        lastUsed:
          type: string
          format: date-time
        status:
          type: string
          enum: [Active, Suspended, Expired]
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        loyaltyCardId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        pointsEarned:
          type: integer
        description:
          type: string
        timestamp:
          type: string
          format: date-time

    # Pagination response
...
security:
  - bearerAuth: []
paths:
...
  /api/loyaltyCards/mine:
    get:
      tags:
        - Loyalty Cards
      summary: Get customer's loyalty cards
      description: Returns all loyalty cards owned by the authenticated customer
      responses:
        '200':
          description: Successfully retrieved loyalty cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoyaltyCard'
        '401':
          description: Unauthorized
  /api/loyaltyCards/{id}:
    get:
      tags:
        - Loyalty Cards
      summary: Get loyalty card by ID
      description: Returns details of a specific loyalty card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved loyalty card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyCard'
        '404':
          description: Card not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - card belongs to another customer
  /api/loyaltyCards/{id}/transactions:
    get:
      tags:
        - Loyalty Cards
      summary: Get transactions for a loyalty card
      description: Returns all transactions for a specific loyalty card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
        '404':
          description: Card not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - card belongs to another customer
  /api/loyaltyCards/{id}/qr-code:
    get:
      tags:
        - Loyalty Cards
      summary: Get QR code for a loyalty card
      description: Returns a QR code for a specific loyalty card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved QR code
          content:
            application/json:
              schema:
                type: object
                properties:
                  qrCode:
                    type: string
        '404':
          description: Card not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - card belongs to another customer
  /api/loyaltyCards/nearby-stores:
    get:
      tags:
        - Loyalty Cards
      summary: Get nearby stores
      description: Returns stores near a geographic location
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: radiusKm
          in: query
          schema:
            type: number
            format: double
            default: 10
      responses:
        '200':
          description: Successfully retrieved nearby stores
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    distance:
                      type: number
                      format: double
        '400':
          description: Invalid parameters
        '401':
          description: Unauthorized

  # Transactions (Customer)
...